Add Another Component for Smaller Header
Considering we rely upon H1,H2,etc imports into the UnderLineText component, I was unable to make this component modular (i.e passing in preferred header tag to be used).

With the previous in mind, I opted for a simpler solution, which is the introduction of an Underline Component but using a smaller header tag.